requirements.txt
fastapi[standard]
requests
fastapi-cache2